Renaming calendar.py to calendar_dialog.py to avoid naming conflicts. Also moved LogName to its own file log_name_dialog.py once again.
