+++ pyqso/calendar_dialog.py
@@ -21,7 +21,7 @@ import logging
 import os.path
 
 
-class Calendar:
+class CalendarDialog:
 
     """ Handler for a simple dialog containing a Gtk.Calendar widget. Using this ensures the date is in the correct YYYYMMDD format required by ADIF. """
